webapp with threading backprocessing

+++ sd_app/views.py
@@ -1,22 +1,26 @@
-from flask import Blueprint, render_template, request, flash, jsonify, redirect, url_for, send_file
+from flask import Flask, Blueprint, render_template, request, flash, jsonify, redirect, url_for, send_file
 import json
+import threading
 from flask_login import login_required, current_user
 from jinja2 import Environment, PackageLoader, select_autoescape
 from .models import Search
 from datetime import datetime as dt
+from sqlalchemy.orm import scoped_session, sessionmaker, Query
+from sqlalchemy import create_engine
 import time
 import pandas as pd
 import sys
 import os
-from . import db
+from . import db, app
 from .constants import keys
-from flask import current_app as app
 sys.path.append('/..')
 from search import search as song_search
 
+flag_bkg = threading.Event()
+
 views = Blueprint('views', __name__)
 input_data = pd.DataFrame(columns=['keyword', 'sp_keyword'])
-input_path = os.path.join(views.root_path, '..', 'input.csv')
+input_path = os.path.join(os.path.dirname(__file__), '..', 'input.csv')
 
 def save_input_data(input_data):
     input_data.to_csv(input_path, header=False)
@@ -28,46 +32,53 @@ def read_data():
     return pd.read_csv(input_path,names=['keyword', 'sp_keyword']).applymap(str)
 
 def add_row(input_data, new_row):
+    def check_kw(kw):
+        flag = False
+        if kw is None or kw == '' or len(kw) == 0: flag = True
+        return flag
+
     #check n of columns
     if len(input_data.columns.values) != 2:
         return False, input_data, "An error ocurred. Try again."
     #check duplicates
     if input_data[(input_data.keyword == new_row['keyword']) & (input_data.sp_keyword == new_row['sp_keyword'])].shape[0] != 0:
         return False, input_data, "Keyword combination already exists."
+    elif check_kw(new_row['keyword']) or check_kw(new_row['sp_keyword']):
+        return False, input_data, "Keywords cannot be empty strings."
     else:
         input_data = pd.concat([input_data, pd.DataFrame(new_row, index=[1])], ignore_index=True)
         return True, input_data, 'Keyword added.'
 
-'''     SCHEDULED NOT WORKING ON PYTHONANYWHERE
-def queue_search(input_data, limit, offset, time_to_complete):
-    queue_path = os.path.join(views.root_path, '..', 'queue.txt')
+def background_search(local_app, local_db, input_data, limit, offset, search_id, Search):
 
-    input_data_dict['input_data'] = input_data.to_dict()
-    input_data_dict['limit'] = limit
-    input_data_dict['offset'] = offset
+    with local_app.app_context():
+        global flag_bkg
 
-    input_data_json = json.dumps(input_data_dict)
+        flag_bkg.set()
+        print("Background search running.")
+        print(f"Limiting results to {limit}. Offset: {offset}")
+        input_data = input_data.rename(columns={'keyword': 'search_term', 'sp_keyword':'keyword'})
 
-    try:
-        with open(queue_path, 'a') as f:
-            f.write(f'{input_data_json}\n')
-        
-        msg = f'Search running in background. Check the search history in about {int(time_to_complete/60)+1} minutes for the download link.'
-        print('Search added to queue')
-    except Exception as e:
-        msg = 'There was an error while writing into the queue. Try again in a few minutes'
-        print('There was an error while writing into the queue')
-        print(e)
-    
-    return msg
-'''
+        filename = song_search(
+            input_data,
+            limit,
+            offset,
+            keys
+        )
+        new_search = Search.query.get(search_id)
+        new_search.csv_path = filename
+        local_db.session.commit()
+        print("Background search completed")
+        flag_bkg.clear()
+        local_db.session.close()
+
 
 @views.route('/', methods=['GET', 'POST'])  
 @views.route('/search', methods=['GET', 'POST']) 
 @login_required 
 def search():
 
-    global input_data
+    global input_data, flag_bkg, app
     filename=''
 
     if request.method == 'GET':
@@ -75,7 +86,6 @@ def search():
         input_data = read_data()
 
     if request.method == 'POST':
-        
         data = dict(request.form)
         
         if data['option'] == 'input_row':
@@ -96,35 +106,39 @@ def search():
 
             try:
                 limit_st = int(data['limit-range-kw-txt']) if data.get('check-limit-kw', None) is not None else -1
-                limit_tot=-1
                 offset = int(data['offset-range-txt']) if data.get('check-offset', None) is not None else 0
                 if offset < 0 or limit_st < -1:
                     flash("Limits must be positive integer numbers. If you don't want to limit or offset the results, uncheck the checkbox", category='error')
                 else:
-                    input_data = read_data()
-                    #time_to_complete = 20*limit_st*len(set(input_data['keyword'].values))
-                    #if time_to_complete <= 330:     #If it takes less than 5 minutes (timeout limit in PA)
+                    if not flag_bkg.is_set():
 
-                    filename = song_search(
-                        input_data.rename(columns={'keyword': 'search_term', 'sp_keyword':'keyword'}),
-                        limit_st,
-                        offset,
-                        keys
-                    )
+                        input_data = read_data()
+                        time_to_complete = 20*limit_st*len(set(input_data['keyword'].values))
+                        
+                        new_search = Search(            #Create search without file path
+                            user = current_user.username,
+                            keywords = input_data.to_json(),
+                            csv_path = "In progress"
+                        )
+                        db.session.add(new_search)
+                        db.session.commit()
+                        search_id = new_search.id
+                        print(search_id)
+                        thread = threading.Thread(target = background_search, kwargs={
+                            'local_app':app,
+                            'local_db': db,
+                            'input_data':input_data,
+                            'limit':limit_st,
+                            'offset':offset,
+                            'search_id':search_id,
+                            'Search': Search
+                        })
 
-                    new_search = Search(
-                        user = current_user.username,
-                        keywords = input_data.to_json(),
-                        csv_path = filename
-                    )
-                    db.session.add(new_search)
-                    db.session.commit()
-                    flash('Search completed.', category='download')
-                    '''
-                    else:           #If the task will take longer, put search on queue
-                        msg = queue_search(input_data.rename(columns={'keyword': 'search_term', 'sp_keyword':'keyword'}), limit_st, offset, time_to_complete)
-                        flash(msg, category='success')
-                    '''
+                        thread.start()
+                        flash(f'Search running in background. Check the search history in about {int(time_to_complete/60)+1} minutes for the download link. Do not close this tab.', category='success')
+                    else:
+                        flash("There's another search running in the background. Try again in a few minutes. Check the search history for completion", category='error')
+                    
             except ValueError:
                 flash("An error happened. Try again.", category='error')
 
@@ -158,32 +172,40 @@ def delete_row():
 
 @views.route('/delete_search', methods=['POST'])  
 @login_required
-def delete_search():
+def delete_search(flash_msg=True, idx = None):
     
-    data = json.loads(request.data)
-    idx = data['idx']
+    if idx == None: idx = json.loads(request.data)['idx']
     search = Search.query.get(idx)
 
-    filepath = os.path.join(views.root_path, 'model_outputs', search.csv_path)
-    print(f"File path to be deleted: {filepath}")
-    os.remove(filepath)
+    if search.csv_path != "In progress":
+        filepath = os.path.join(views.root_path, 'model_outputs', search.csv_path)
+        if os.path.exists(filepath):
+            print(f"File path to be deleted: {filepath}")
+            os.remove(filepath)
 
     db.session.delete(search)
     db.session.commit()
 
-    flash('Search record deleted.', category='error')
+    if flash_msg: flash('Search record deleted.', category='error')
 
     return jsonify({})
 
-@views.route('/clear_input', methods=['POST'])  
+@views.route('/clear', methods=['POST'])  
 @login_required
-def clear_input(flash_msg=True):
+def clear(flash_msg=True, what_to_clear=None):
     global input_data
     
-    input_data = pd.DataFrame(columns=['keyword', 'sp_keyword'])
+    if what_to_clear is None: what_to_clear = json.loads(request.data)['what_to_clear']
+    if what_to_clear == 'input':
+        input_data = pd.DataFrame(columns=['keyword', 'sp_keyword'])
 
-    save_input_data(input_data)
-    if flash_msg: flash('Input data cleared.', category='error')
+        save_input_data(input_data)
+        if flash_msg: flash('Input data cleared.', category='error')
+    
+    elif what_to_clear == 'history':
+        for search in Search.query.all():
+            delete_search(flash_msg=False, idx = search.id)
+        flash('Search history cleared.', category='error')
 
     return jsonify({})
 
@@ -192,7 +214,7 @@ def repeat_search():
     global input_data
     search_input = json.loads(request.data)['keyword']
 
-    clear_input(flash_msg=False)
+    clear(flash_msg=False, what_to_clear='input')
     input_data = pd.DataFrame(search_input)
 
     save_input_data(input_data)
